SCC-2109
This update reorganizes the is-research-service to remove the Lambda layer format. This make this much easier to maintain and invoke because testing changes no longer requires generating a new version of the layer to validate logic changes. This also allows for a cleaner repo organization that makes locating source files easier.